Menu icons and beginnings of terrain
I created a new prefab for our terrain which will hold all of the objects. i created a mesh material for the pathways and also added random grass that spawns on the platform. I added the visuals for all of the buttons and made the play button, the menu button and the close button functional.